+++ src/main/java/numble/challenge/NumbleChallengeApplication.java
@@ -1,4 +1,4 @@
-package Numble.NumbleChallenge;
+package numble.challenge;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
