+++ src/main/java/numble/challenge/order/service/OrderService.java
@@ -1,12 +1,14 @@
-package Numble.NumbleChallenge.domain.order.service;
+package numble.challenge.order.service;
 
-import Numble.NumbleChallenge.domain.order.Order;
+import numble.challenge.domain.model.entity.Order;
 
 import java.util.List;
 
 public interface OrderService {
     Order createOrder(Long memberId, Long itemId, int count);
+
     void cancelOrder(Long orderId);
+
     List<Order> findAllOrder();
 
     List<Order> findOrderByMember(Long memberId);
