+++ src/test/java/numble/challenge/NumbleChallengeApplicationTests.java
@@ -1,4 +1,4 @@
-package Numble.NumbleChallenge;
+package numble.challenge;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
